requirements.txt
mysqlclient==2.1.1
asgiref==3.7.2
Django==3.2.20
django-cors-headers==4.1.0
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.1
psycopg2==2.9.7
python-dotenv==0.21.1
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
whitenoise==6.5.0
geopy==2.3.0
selenium==4.10.0
Unidecode==1.3.6
requests==2.26.0
